Ensure volume region matches instance
Some instance types are not available in the eu-west-2 region yet, which
is our favoured zone. Fixing the volume region on the volume to match
the instance makes it easy to set the region as a variable in main.tf or
as an argument with `--var` at the command line.